move choicecard to packages
