+++ packages/@guardian/src-choice-card/ChoiceCard.stories.tsx
@@ -1,6 +1,6 @@
 import { SvgCamera } from '@guardian/src-icons';
-import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
-import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
+import { Story } from '../../../lib/@types/storybook-emotion-10-fixes';
+import { asChromaticStory, asPlayground } from '../../../lib/story-intents';
 import { ChoiceCard, ChoiceCardProps } from './ChoiceCard';
 
 export default {
