+++ packages/@guardian/src-choice-card/ChoiceCardGroup.stories.tsx
@@ -1,8 +1,8 @@
 import { ThemeProvider } from '@emotion/react';
 import { choiceCardDefault } from '@guardian/src-foundations/themes';
 import React, { useState } from 'react';
-import { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
-import { asChromaticStory, asPlayground } from '../../../../lib/story-intents';
+import { Story } from '../../../lib/@types/storybook-emotion-10-fixes';
+import { asChromaticStory, asPlayground } from '../../../lib/story-intents';
 import { ChoiceCard } from './ChoiceCard';
 import ChoiceCardStories from './ChoiceCard.stories';
 import { ChoiceCardGroup, ChoiceCardGroupProps } from './ChoiceCardGroup';
